ex8
ex8 finished.

Fixed checkbox Issue

current Issues:

Can't select in progress in dropdown. I added a classname to each Chips element but it does not seem to allow me to find it if the className includes space like "in Progress"

Need to check if I can ge the selectedValue from the dropdown in order to do a proper check.
Sometimes it does not delete todos properly from server but does delete from state